@@ -4,11 +4,11 @@
   "scripts": {
     "ng": "ng",
     "prestart": "npm run build:copy-changelog",
-    "start": "ng serve -o",
+    "start": "ng serve -o --hmr false",
     "start:prod": "npm run build && node ./projects/elements-server/server.js",
     "lint": "ng lint",
     "test": "ng lint && npm run format:test && npm run test:lib",
-    "test:lib": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --config projects/elements/jest.config.mjs",
+    "test:lib": "node --experimental-vm-modules --no-warnings node_modules/jest/bin/jest.js --config projects/elements/jest.config.ts",
     "watch": "ng test --watch",
     "prebuild": "",
     "build": "ng build elements-demo && ng build elements",
@@ -47,34 +47,34 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^18.1.0",
-    "@angular/cdk": "^18.1.0",
-    "@angular/common": "^18.1.0",
-    "@angular/compiler": "^18.1.0",
-    "@angular/core": "^18.1.0",
-    "@angular/forms": "^18.1.0",
-    "@angular/material": "^18.1.0",
-    "@angular/platform-browser": "^18.1.0",
-    "@angular/platform-browser-dynamic": "^18.1.0",
-    "@angular/pwa": "^18.1.0",
-    "@angular/router": "^18.1.0",
-    "@angular/service-worker": "^18.1.0",
+    "@angular/animations": "^19.2.8",
+    "@angular/cdk": "^19.2.11",
+    "@angular/common": "^19.2.8",
+    "@angular/compiler": "^19.2.8",
+    "@angular/core": "^19.2.8",
+    "@angular/forms": "^19.2.8",
+    "@angular/material": "^19.2.11",
+    "@angular/platform-browser": "^19.2.8",
+    "@angular/platform-browser-dynamic": "^19.2.8",
+    "@angular/pwa": "^19.2.9",
+    "@angular/router": "^19.2.8",
+    "@angular/service-worker": "^19.2.8",
     "highlight.js": "^11.3.1",
     "ngx-highlightjs": "^12.0.0",
-    "ngx-markdown": "^18.0.0",
+    "ngx-markdown": "^19.1.1",
     "rxjs": "^7.5.0",
     "tslib": "^2.3.1",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.1.0",
-    "@angular-eslint/builder": "^18.1.0",
-    "@angular-eslint/eslint-plugin": "^18.1.0",
-    "@angular-eslint/eslint-plugin-template": "^18.1.0",
-    "@angular-eslint/schematics": "^18.1.0",
-    "@angular-eslint/template-parser": "^18.1.0",
-    "@angular/cli": "^18.1.0",
-    "@angular/compiler-cli": "^18.1.0",
+    "@angular-eslint/builder": "^19.3.0",
+    "@angular-eslint/eslint-plugin": "^19.3.0",
+    "@angular-eslint/eslint-plugin-template": "^19.3.0",
+    "@angular-eslint/schematics": "^19.3.0",
+    "@angular-eslint/template-parser": "^19.3.0",
+    "@angular/build": "^19.2.9",
+    "@angular/cli": "^19.2.9",
+    "@angular/compiler-cli": "^19.2.8",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@jest/globals": "^29.7.0",
@@ -90,8 +90,8 @@
     "express": "^4.17.1",
     "husky": "^4.3.8",
     "jest": "^29.7.0",
-    "jest-preset-angular": "^14.1.1",
-    "ng-packagr": "^18.1.0",
+    "jest-preset-angular": "^14.5.5",
+    "ng-packagr": "^19.2.2",
     "prettier": "^3.3.2",
     "pretty-quick": "^4.0.0",
     "raw-loader": "^4.0.2",
