@@ -1,4 +1,6 @@
-import 'jest-preset-angular/setup-jest.mjs';
+import { setupZoneTestEnv } from 'jest-preset-angular/setup-env/zone/index.mjs';
+
+setupZoneTestEnv();
 
 declare const ngDevMode: boolean;
 
